Replace [Fact] with [Theory]
